+++ Wrappers/Python/test/test_algorithms.py
@@ -552,94 +552,18 @@ class TestAlgorithms(CCPiTestClass):
         self.assertLess(rmse, 4.2e-4)
 
 
-    def test_exception_initial_CGLS(self):
-        ig = ImageGeometry(10,2)
-        numpy.random.seed(2)
-        initial = ig.allocate(0.)
-        b = ig.allocate('random')
-        identity = IdentityOperator(ig)
-        
-        try:
-            alg = CGLS(initial=initial, operator=identity, data=b, x_init=initial)
-            assert False
-        except ValueError as ve:
-            logging.info(str(ve))
-            assert True
-
-            
-    def test_exception_initial_FISTA(self):
-        ig = ImageGeometry(127,139,149)
-        initial = ig.allocate()
-        b = initial.copy()
-        # fill with random numbers
-        b.fill(numpy.random.random(initial.shape))
-        initial = ig.allocate(ImageGeometry.RANDOM)
-        identity = IdentityOperator(ig)
-        
-        norm2sq = OperatorCompositionFunction(L2NormSquared(b=b), identity)
-        opt = {'tol': 1e-4, 'memopt':False}
-        logging.info("initial objective {}".format(norm2sq(initial)))
-        try:
-            alg = FISTA(initial=initial, f=norm2sq, g=ZeroFunction(), x_init=initial)
-            assert False
-        except ValueError as ve:
-            logging.info(str(ve))
-            assert True
             
 
-    def test_exception_initial_GD(self):
-        ig = ImageGeometry(127,139,149)
-        initial = ig.allocate()
-        b = initial.copy()
-        # fill with random numbers
-        b.fill(numpy.random.random(initial.shape))
-        initial = ig.allocate(ImageGeometry.RANDOM)
-        identity = IdentityOperator(ig)
-        
-        norm2sq = OperatorCompositionFunction(L2NormSquared(b=b), identity)
-        opt = {'tol': 1e-4, 'memopt':False}
-        logging.info("initial objective {}".format(norm2sq(initial)))
-        try:
-            alg = GD(initial=initial, objective_function=norm2sq, x_init=initial)
-            assert False
-        except ValueError as ve:
-            logging.info(str(ve))
-            assert True
 
 
-    def test_exception_initial_LADMM(self):
-        ig = ImageGeometry(10,10)
-        # K = Identity(ig)
-        # b = ig.allocate(0)
-        # f = LeastSquares(K, b)
-        # g = IndicatorBox(lower=0)
-        initial = ig.allocate(1)
-        try:
-            algo = LADMM(initial = initial, x_init=initial)
-            assert False
-        except ValueError as ve:
-            logging.info(str(ve))
-            assert True
 
 
-    def test_exception_initial_PDHG(self):
-        initial = 1
-        try:
-            algo = PDHG(initial = initial, x_init=initial, f=None, g=None, operator=None)
-            assert False
-        except ValueError as ve:
-            logging.info(str(ve))
-            assert True
 
 
-    def test_exception_initial_SPDHG(self):
-        initial = 1
-        try:
-            algo = SPDHG(initial = initial, x_init=initial)
-            assert False
-        except ValueError as ve:
-            logging.info(str(ve))
-            assert True
+
+
+
+
             
 
 class TestSIRT(unittest.TestCase):
@@ -705,14 +629,6 @@ class TestSIRT(unittest.TestCase):
         np.testing.assert_almost_equal(alg.solution.max(), 0.3)  
         np.testing.assert_almost_equal(alg.solution.min(), 0.1) 
         
-
-    def test_exception_initial_SIRT(self):
-        try:
-            alg = SIRT(initial=self.initial2, operator=self.A2, data=self.b2, x_init=self.initial2)
-            assert False
-        except ValueError as ve:
-            logging.info(str(ve))
-            assert True     
 
 
     def test_SIRT_nan_inf_values(self):